requirements.txt
Django==2.1.2
djangorestframework==3.9.0
Pillow==5.3.0
Wand==0.4.4